requirement.txt
numpy
plotly_express
dash
dash_core_components
dash_html_components
folium
pandas